+++ release/e/eo_plus/HISTORY.md
@@ -1,6 +1,15 @@
 Esperanto Plus Change History
 =============================
 
+3.0 (2024.01.30)
+----------------
+- Removed the '-et-' key from the touch layout
+- Added the 'radiko' to the touch layout
+- Updated layout code to handle the 'radiko' key
+- Fixed letters not uppercase when caps lock
+- Updated welcome and help files
+- Added the 'mns' shortcut for 'Mi ne scias'
+
 2.0 (2024-01-12)
 ----------------
 - Modified the help file